Move '/syncer/forms/*' forms to '/syncer/uploader/forms/*'
